# java/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Polymorpic Swagger",
    "description": "",
    "contact": {
      "name": "Dev",
      "url": "https://somecompany.com",
      "email": "developers@somecompany.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ProductModel"
                      },
                      {
                        "$ref": "#/components/schemas/ProductModelV2"
                      }
                    ]
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ProductModel"
                      },
                      {
                        "$ref": "#/components/schemas/ProductModelV2"
                      }
                    ]
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ProductModel"
                      },
                      {
                        "$ref": "#/components/schemas/ProductModelV2"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Add product to available products",
        "description": "You could for instance add products using\r\n```JSON\r\n{\r\n  \"cost\":124, \"name\": \"tea\"\r\n}\r\n```",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AddProduct"
                  },
                  {
                    "$ref": "#/components/schemas/AddProduct2"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AddProduct"
                  },
                  {
                    "$ref": "#/components/schemas/AddProduct2"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AddProduct"
                  },
                  {
                    "$ref": "#/components/schemas/AddProduct2"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The product was added successfully"
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "example": {"cost":"The field Cost must be between 0.1 and 1000000.","name":"The Name field is required."}
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "example": {"cost":"The field Cost must be between 0.1 and 1000000.","name":"The Name field is required."}
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "example": {"cost":"The field Cost must be between 0.1 and 1000000.","name":"The Name field is required."}
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProductModel"
                    },
                    {
                      "$ref": "#/components/schemas/ProductModelV2"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProductModel"
                    },
                    {
                      "$ref": "#/components/schemas/ProductModelV2"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProductModel"
                    },
                    {
                      "$ref": "#/components/schemas/ProductModelV2"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddProduct": {
        "required": [
          "cost",
          "name",
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "cost": {
            "maximum": 1000000,
            "minimum": 0.1,
            "type": "number",
            "description": "",
            "format": "currency"
          },
          "name": {
            "maxLength": 60,
            "minLength": 3,
            "type": "string",
            "description": ""
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "version",
          "mapping": {
            "AddProduct": "#/components/schemas/AddProduct",
            "AddProduct2": "#/components/schemas/AddProduct2"
          }
        }
      },
      "AddProduct2": {
        "required": [
          "properties"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/AddProduct"
          }
        ],
        "properties": {
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": ""
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProductModel": {
        "required": [
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "format": "float"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "version",
          "mapping": {
            "ProductModel": "#/components/schemas/ProductModel",
            "ProductModelV2": "#/components/schemas/ProductModelV2"
          }
        }
      },
      "ProductModelV2": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ProductModel"
          }
        ],
        "properties": {
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}